+++ public/install.php
@@ -516,7 +516,6 @@ function pythonHealth(): array
 	]);
 	$response = curl_exec($ch);
 	$httpCode = curl_getinfo($ch, CURLINFO_HTTP_CODE);
-	curl_close($ch);
 
 	if ($httpCode === 200 && $response) {
 		$json = json_decode($response, true);
@@ -843,7 +842,6 @@ function handleFinalize(): never
 		curl_setopt_array($ch, [CURLOPT_RETURNTRANSFER => true, CURLOPT_TIMEOUT => 3, CURLOPT_CONNECTTIMEOUT => 2]);
 		$response = curl_exec($ch);
 		$pythonReloaded = curl_getinfo($ch, CURLINFO_HTTP_CODE) === 200 && $response !== false;
-		curl_close($ch);
 	}
 
 	unset(